// frontend/front_pfe_standalone-houssem/src/app/services/models/projet-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProductBacklog } from '../models/product-backlog';
import { User } from '../models/user';
export interface ProjetResponse {
  actif?: boolean;
  dateDebut?: string;
  description?: string;
  equipe?: Array<User>;
  id?: number;
  nom?: string;
  productBacklog?: ProductBacklog;
  productOwner?: User;
  scrumMaster?: User;
  statut?: 'EN_COURS' | 'TERMINE' | 'EN_PAUSE' | 'A_FAIRE';
}