// frontend/front_pfe_standalone-houssem/src/app/services/models/projet-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { User } from '../models/user';
export interface ProjetRequest {
  actif?: boolean;
  dateDebut?: string;
  description: string;
  equipe?: Array<User>;
  equipeIds?: Array<number>;
  id?: number;
  nom: string;
  productOwner?: User;
  productOwnerId?: number;
  scrumMaster?: User;
  scrumMasterId?: number;
}